HAAR-1930: merged SaveClientService.kt to ClientsService.kt
